+++ custom_components/supernotify/__init__.py
@@ -214,6 +214,8 @@ OPTION_TARGET_CATEGORIES = "target_categories"
 OPTION_UNIQUE_TARGETS = "unique_targets"
 OPTION_TARGET_INCLUDE_RE = "target_include_re"
 OPTION_CHIME_ALIASES = "chime_aliases"
+OPTION_DATA_KEYS_INCLUDE_RE = "data_keys_include_re"
+OPTION_DATA_KEYS_EXCLUDE_RE = "data_keys_exclude_re"
 
 RE_DEVICE_ID = r"^[0-9a-f]{32}$"