+++ custom_components/supernotify/notify.py
@@ -64,7 +64,7 @@ from .context import Context
 from .delivery import DeliveryRegistry
 from .hass_api import HomeAssistantAPI
 from .media_grab import MediaStorage
-from .model import ConditionVariables
+from .model import ConditionVariables, SuppressionReason
 from .notification import Notification
 from .people import PeopleRegistry, Recipient
 from .scenario import ScenarioRegistry
@@ -442,7 +442,20 @@ class SupernotifyAction(BaseNotificationService):
             elif notification.errored:
                 _LOGGER.error("SUPERNOTIFY Failed to deliver %s, error count %s", notification.id, notification.errored)
             else:
-                _LOGGER.warning("SUPERNOTIFY No deliveries made for  %s", notification.id)
+                if notification.undelivered_envelopes:
+                    codes: list[SuppressionReason] = list({
+                        envelope.skip_reason if envelope.skip_reason else SuppressionReason.UNKNOWN
+                        for envelope in notification.undelivered_envelopes
+                    })
+                    reason: str = ",".join(str(code) for code in codes)
+                    problem: bool = codes != [SuppressionReason.DUPE]
+                else:
+                    problem = True
+                    reason = "No delivery envelopes generated"
+                if problem:
+                    _LOGGER.warning("SUPERNOTIFY No deliveries made for %s: %s", notification.id, reason)
+                else:
+                    _LOGGER.debug("SUPERNOTIFY Deliveries suppressed for %s: %s", notification.id, reason)
 
         except Exception as err:
             # fault barrier of last resort, integration failures should be caught within envelope delivery